+++ src/components/Form/MucRadioButtonGroup.vue
@@ -16,7 +16,7 @@
 <script setup lang="ts">
 import { provide, readonly, toRef } from "vue";
 
-import { RadioButtonGroupKey, RadioButtonValueTypes } from "./RadioButtonTypes";
+import { RadioButtonGroupKey, RadioButtonValueTypes } from "./MucRadioButtonTypes";
 
 /**
  * exposed two-way binding of the currently selected radiobuttons-value
